+++ database/migrations/2017_05_01_141528_DeleteDownloadTable.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class DeleteDownloadTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::dropIfExists('downloads');
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::create('downloads', function (Blueprint $table) {
+            $table->increments('id');
+            $table->char('token', 36)->unique();
+            $table->char('server', 36);
+            $table->text('path');
+            $table->timestamps();
+        });
+    }
+}