+++ database/migrations/2017_05_01_141559_DeleteNodeConfigurationTable.php
@@ -1,0 +1,37 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class DeleteNodeConfigurationTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::dropIfExists('node_configuration_tokens');
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::create('node_configuration_tokens', function (Blueprint $table) {
+            $table->increments('id');
+            $table->char('token', 32);
+            $table->unsignedInteger('node_id');
+            $table->timestamps();
+        });
+
+        Schema::table('node_configuration_tokens', function (Blueprint $table) {
+            $table->foreign('node_id')->references('id')->on('nodes');
+        });
+    }
+}